+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Use AWS SDK V3 for `s3Submit` service
+
 ## [5.1.0] - 2023-07-03
 
 ### Changed